assignments/lab6/requirements.txt
googleapis-common-protos
google-cloud-storage